test_server: work with offsets work when queried twice
There were 2 bugs, first when I added support for paged results I did
nothing in the test_server. So it just gave you the same bugs infinite
number of times

Second, on a query the testClient stored the bugs and bugList read from
disk. On subsequent calls it would check if the bugs in the map were set
and if so shortcircuit reading from disk. The problem is that the list
of bugs are in a struct, not a map AND the testClient.Search was pass by
value, not reference. So query was able to update the map, but not the
struct. Pass testClient by reference instead.